+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 5.2.6
 -----
 
+* `Support for Python 3.6 and NumPy<1.17 has been dropped <https://numpy.org/neps/nep-0029-deprecation_policy.html>`_
 * Windows installers are now built with pynsist/NSIS instead of PyInstaller (#15).
 
 5.2.5